requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot[ext]
pydantic
pydantic-settings
aiohttp
aiosqlite
aiofiles
requests
python-dotenv
msal
